Refactored CodeUniStream#matchOne to return a Bool instead.
Since translating the code from Dart to Swift, I noticed in the return types the difference among methods that were returning Int, UInt8? and Bool. In Dart, they were either Int or Bool only, so I hadn't noticed as much difference there.

In Swift, the 3 different types made me realize that it was time to change some of the match methods could return Bool instead, simplifying the code. And while in Dart I was trying to keep the match methods returning about the same type for consistency, in Swift they were more diverse and the compiler was always checking for them anyway.